chore: remove AvatarDomain
