--- Editor/Domain/AvatarDomain.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a7f535cb6156d645aeda85ca3a58c88
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
